feat: simplify and enhance flight types, removing legacy definitions
