--- webapp/src/types/flightTypes.ts
@@ -1,28 +1,0 @@
-// Flight query interface
-export interface FlightQuery {
-  origin: string;
-  dest: string;
-  depDate: string;
-  retDate?: string;
-  numAdults?: number;
-  numChildren?: number;
-  numInfants?: number;
-  cabinClass?: string;
-}
-
-// Flight result interface
-export interface FlightResult {
-  price: string;
-  duration: string;
-  stops: number;
-  airline: string;
-  departure: string;
-  arrival: string;
-  origin: string;
-  destination: string;
-  departureDate: string;
-  returnDate?: string;
-  layoverAirports?: string[];
-  layoverDurations?: string[];
-  cabinClass?: string;
-}
